feat: take logic out from http controller
