+++ crates/server/src/api/routes/quests/get_quest.rs
@@ -6,7 +6,7 @@ use quests_definitions::quests::Quest;
 use serde::Serialize;
 use utoipa::ToSchema;
 
-use crate::api::routes::errors::QuestError;
+use crate::logic::quests::QuestError;
 
 use super::types::ToQuest;
