+++ crates/server/src/api/routes/quests/types.rs
@@ -1,7 +1,7 @@
 use quests_db::core::definitions::{StoredQuest, UpdateQuest};
 use quests_definitions::quests::Quest;
 
-use crate::api::routes::errors::QuestError;
+use crate::logic::quests::QuestError;
 
 pub trait ToQuest {
     fn to_quest(&self) -> Result<Quest, QuestError>;
