+++ crates/server/src/logic/mod.rs
@@ -1,0 +1,2 @@
+pub mod events;
+pub mod quests;
